Move rule and start_state codegen to module
